+++ src/nodes/ScrollLayerExt.cpp
@@ -173,7 +173,7 @@ void ScrollLayerExt::fixTouchPrio() {
     if (auto delegate = typeinfo_cast<CCTouchDelegate*>(this)) {
         if (auto handler = CCTouchDispatcher::get()->findHandler(delegate)) {
             Loader::get()->queueInMainThread([this, handler, delegate, oldThis]() {
-                if (oldThis != nullptr) {
+                if (oldThis != nullptr && handler != nullptr && delegate != nullptr) {
                     if (auto dispatcher = CCTouchDispatcher::get()) {
                         dispatcher->setPriority(handler->m_nPriority - 2, delegate);
                     }